pyproject.toml
[tool.poetry]
name = "amphisbaena"
version = "0.1.0"
description = ""
authors = ["Jeremy Fisher <jeremyadamsfisher@gmail.com>"]
packages = [{include = "amphisbaena"}]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.12.1"
torchvision = "^0.13.1"
pytorch-lightning = "^1.7.7"
pytorch-lightning-bolts = "^0.3.2"
loguru = "^0.6.0"
scipy = "1.6"
numpy = "1.21"
matplotlib = "3.5"
sklearn = "^0.0"
gym = "^0.26.2"
click = "^8.1.3"
wandb = "^0.13.4"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
cli = "amphisbaena.cli:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

